Move films container to component
