+++ src/main.js
@@ -1,12 +1,9 @@
 import Profile from './components/profile';
 import Menu from './components/menu';
 import Sort from './components/sort';
-import FilmsList from './components/films-list';
+import Films from './components/films';
 import Details from './components/details';
 import Statistics from './components/statistics';
-
-const QUANTITY_CARDS_UPCOMING = 5;
-const QUANTITY_CARDS_TOP = 2;
 
 const siteHeaderElem = document.querySelector(`.header`);
 const siteMainElem = document.querySelector(`.main`);
@@ -15,41 +12,9 @@ const siteFooterElem = document.querySelector(`.footer`);
 const profile = new Profile();
 const menu = new Menu();
 const sort = new Sort();
+const films = new Films();
 const details = new Details();
 const statistics = new Statistics();
-
-const filmsSectionsData = [
-  {
-    type: 'upcoming',
-    title: 'All movies. Upcoming',
-    quantity: QUANTITY_CARDS_UPCOMING
-  },
-  {
-    type: 'extra',
-    title: 'Top rated',
-    quantity: QUANTITY_CARDS_TOP
-  },
-  {
-    type: 'extra',
-    title: 'Most commented',
-    quantity: QUANTITY_CARDS_TOP
-  }
-];
-
-const getFilmsTmpl = () => {
-  const filmsSections = filmsSectionsData
-    .reduce((prev, data) => {
-      const filmsSection = new FilmsList(data);
-      return prev + filmsSection.getTmpl();
-    } ,'');
-
-  return (
-    `<section class="films">
-      ${filmsSections}
-    </section>
-    `
-  );
-};
 
 const render = (container, template) => {
   container.insertAdjacentHTML(`beforeend`, template);
@@ -59,7 +24,7 @@ render(siteHeaderElem, profile.getTmpl());
 
 render(siteMainElem, menu.getTmpl());
 render(siteMainElem, sort.getTmpl());
-render(siteMainElem, getFilmsTmpl());
+render(siteMainElem, films.getTmpl());
 
 render(siteFooterElem, statistics.getTmpl());
